Shift around some functions to avoid circular imports in json_rewrite.py
