+++ elifetools/json_rewrite.py
@@ -3,8 +3,9 @@
 from collections import OrderedDict
 from six import iteritems
 
-import elifetools.parseJATS 
+import elifetools.rawJATS
 import elifetools.utils
+import elifetools.utils_html
 
 def rewrite_json(rewrite_type, soup, json_content):
     """
@@ -13,15 +14,19 @@ def rewrite_json(rewrite_type, soup, json_content):
     """
     if not soup:
         return json_content
-    if not elifetools.parseJATS.doi(soup) or not elifetools.parseJATS.journal_id(soup):
+    if not elifetools.rawJATS.doi(soup) or not elifetools.rawJATS.journal_id(soup):
         return json_content
 
     # Hook only onto elife articles for rewriting currently
-    if elifetools.parseJATS.journal_id(soup).lower() == "elife":
-        function_name = rewrite_function_name(elifetools.parseJATS.journal_id(soup), rewrite_type)
+    journal_id_tag = elifetools.rawJATS.journal_id(soup)
+    doi_tag = elifetools.rawJATS.doi(soup)
+    journal_id = elifetools.utils.node_text(journal_id_tag)
+    doi = elifetools.utils.doi_uri_to_doi(elifetools.utils.node_text(doi_tag))
+    if journal_id.lower() == "elife":
+        function_name = rewrite_function_name(journal_id, rewrite_type)
         if function_name:
             try:
-                json_content = globals()[function_name](json_content, elifetools.parseJATS.doi(soup))
+                json_content = globals()[function_name](json_content, doi)
             except KeyError:
                 pass
     return json_content
@@ -366,9 +371,9 @@ def elife_references_rewrite_json():
         ref_json[id][author_type] = []
         for ref_author in authors:
             if  "collab" in ref_author:
-                author_json = elifetools.parseJATS.references_author_collab(ref_author)
+                author_json = elifetools.utils_html.references_author_collab(ref_author)
             else:
-                author_json = elifetools.parseJATS.references_author_person(ref_author)
+                author_json = elifetools.utils.references_author_person(ref_author)
             if author_json:
                 ref_json[id][author_type].append(author_json)
         # Add to json array, and do not verwrite existing rule of a specific bib id (if present)
